SERVER-48164 Override install-action in individual tests
In a previous commit, install-action was set to hardlink for all tests
to reduce resource usage. However, that option doesn't work for one of
the tests. The reasons why it doesn't work are beyond the scope of this
commit, but it probably won't be the last time we need to change the
install mode for a single test, either.

The global install-action will now default to hardlink but can also be
overridden in a test where it is incompatible or where we have reasons
to use a different install-action.

(cherry picked from commit ece6d35fbe5337fc9ce22094367905e5cab712ff)