SERVER-48933 Fix fork-after-new-thread death test failures with TSAN
The death tests fork() a new thread in a multithreaded test harness so
we can test thread death behavior. Unfortunately TSAN does not support
fork() after spawning a thread and will kill the thread by default,
causing a test failure. This sets an unsupported TSAN flag that disables
the behavior, allowing the tests to proceed. There is currently no means
to limit the change to only the death tests, but the change only applies
to the death tests for now.