SERVER-39191 Optimize count queries on sharded collections
When counts were made orphan-aware the addition of the SHARD_FILTER
stage started preventing the COUNT_SCAN optimization on all sharded
collections, leading to ~5x performance degradation due to the necessity
of IXSCAN (and possibly FETCH).

However, there can not be any orphans if the shard key is fully
specified: A mongodb instance will never have orphans for chunk ranges
it owns and mongos will send a query with a full shard key to one
replicaset only.

This allows stripping the SHARDING_FILTER stage for cases where the
query specifies the full shard key, regaining the lost performance.