SERVER-46199 Get rid of the BM_IncrementalRefreshWithNoChange benchmark
This test is incompatible with making the RoutingTableHistory movable,
done under 3df9c61b7747058cd84edd7a875b1524db57dd1d. In order to
preserve the same performance characteristic, the check for whether any
changes have happened to the routing info was pulled at a higher level,
in the CatalogCache.