SERVER-51058 Use rawMongoProgramOutput carefully in log_remote_op_wait.js
On the v4.2 branch, log_remote_op_wait.js uses the rawMongoProgramOutput()
function to assert that a query produces the right log line. But
rawMongoProgramOutput() doesn't necessarily include the log line,
even though the server did write it, and the client did receive the
response. The problem is that the command response and the log line
happen over two different channels (the network, vs a background thread
reading from a pipe).

It looks like other jstests fix this by shutting down the server before
checking rawMongoProgramOutput(); that should work here too.

This only affects the v4.2 branch. Originally log_remote_op_wait.js used
the getLog command, but when backporting the test to 4.2 we switched
one assertion to rawMongoProgramOutput to overcome the log-line length
limit of getLog.