SERVER-40361 Don't store debug info once plan cache size grows large
Introduces a new setParameter,
'internalQueryCacheMaxSizeBytesBeforeStripDebugInfo'. When
the cumulative size of a mongod's plan caches exceeds this
threshold, additional plan cache entries are stored without
any debug info. This should help to prevent problems where
the plan caches collectively consume too much memory.

The default setting of the parameter is 0.5 GB, but it can
be configured by the operator at startup or at runtime.