SERVER-52690 Run ReshardingCollectionCloner on a TaskExecutor.
Restructures the ReshardingCollectionCloner class to run on a separate
thread and returned a mongo::Future readied when the cloning procedure
has completed successfully or failed with an error.

Also enables the ReshardingCollectionCloner to insert documents in
batches of size > 1 in a single storage transaction. The
reshardingCollectionClonerBatchSizeInBytes server parameter can be used
to control the size of the insert batch.