SERVER-52610 Ensure RPM install roots can be relocated
We were not using directory macros in the RPMs, which meant that the
installation root could not be changed. When we tried, we got an
incomplete install, with some of it going to the new prefix and some
going to the host system. Here, we follow RPM packaging guidelines a bit
better and use named directory macros instead of explicit paths.

(cherry picked from commit 9aba0cac770d87d36cf0f0d19a1be06e73312221)