SERVER-51839 Add error context when document validator with $expr fails by throwing an exception
Document validators which use $expr will now generate detailed errors
when $expr throws. When the validationAction is set to error, the write
will fail with a DocumentValidationError instead of the exception's
error code. When the validationAction is set to warn, the write will be
accepted and a detailed error will be logged.