SERVER-52788 Rename measureOperationResourceConsumption feature flag to profileOperationResourceConsumptionMetrics
This renames the "measureOperationResourceConsumption" feature flag to
"profileOperationResourceConsumptionMetrics". It is a default-off flag that
enables a subset of metrics collection that can be profiled and reported in the
slow query logs. The "aggregateOperationResourceConsumptionMetrics" flag, in
addition to enabling global aggregation and the use of $operationMetrics, also
enables profiling and slow query logging covered by the "profile.." flag.