SERVER-53579 Update mypy version
The version of mypy we were depending on was very old and was causing a
version conflict on typed-ast with another module. On newer pip versions,
the result was a *very* slow process of walking back compatible versions
of multiple packages. This change updates mypy to the latest version
available.