SERVER-57492 Side-write table should not write more than KeyString
This fixes a bug where extra data in the form of uninitialized memory was being
written to the index builds side writes table. This memory is never observed
and only takes up space in the table.

This commit also renames BufBuilder::getSize() to capacity() to conform
to the STL convention and to prevent similar bugs in the future.

(cherry picked from commit ba81849df916f6c417585c6cd51e6508be5001fe)