SERVER-56486 Improve SBE multiplanner
This patch introduces a number of improvements to the SBE multiplanner.
In particular:
- We no longer run plans in round-robin fashion (i.e. switch between
plans until each has returned one result, and so on). Instead, each
plan runs until it finishes its trial period.
- The number of reads is progressively decreased. More precisely, the
number of reads used by a plan during is trial period is used as the
maximum number of reads for subsequent plans.
- Non blocking plans are run before blocking ones. This is a heuristic
that limits the amount of work done by a blocking plan as each blocking
plan will now be bounded by the smallest number of reads used by a non
blocking plan during its trial period.