SERVER-58313 Fix package testing for SLES 12 on v4.2 and v4.0
SERVER-55460 attempted to fix SLES 12 package testing, but it continued
to fail on v4.2 and v4.0. We thought it was due to a recent SLES 12
package testing image change, but that wasn't actually the case. It was
actually because the `-tools-extra` package doesn't exist in those
version, and we should have been testing the `-tools` package instead.
This commit fixes the previous backport so the correct package is
tested.