SERVER-58578 Timeseries bucket compression
All fields under data will be compressed using BSON Binary Subtype 7.
Measurements are sorted by time field.

Buckets are compressed in a separate update operation when the
BucketCatalog is closing the bucket and no further writes to it will be
performed.

Compression operation are transparant for the user, if it fails for any
reason the bucket will be left uncompressed.