SERVER-60097 Add method to wait for topology state
The TenantSplitDonorService needs to wait for three non-voting
replica set nodes to complete an initial sync before the reconfig
which will establish them as the recipient slice. This patch
introduces a new method `awaitTopologyState` which allows us to
wait for the local replica set to meet some provided condition.
Additionally, a new "waiter list" was introduced since the
existing `WaiterList` is primarily focused on waiting for opTimes
which is unrelated to topology state changes.