SERVER-58580 Decoding support for compressed subobjects in BSONColumn
BSONColumn has been refactored to put the decoding state into a DecodingState class.

A new control byte in the data stream has been added for entering "interleaved" mode where sub-elements are compressed separately.

One DecodingState is set up for each scalar element in the reference interleaved Object. When decompressing we extract one value from each state to build the full BSONObj to return. We always scan in a fixed order.

Control bytes in the binary stream corresponds to the DecodingState that needs to read the next value.