Fix "mongod --shutdown" to check size of "mongod.lock" instead of existence
In a07dc78e49c53de539dff92748334cd7fbe2e2a9, the code for `mongod --shutdown` was refactored, and appears to have inadvertently changed from `while (boost::filesystem::exists(procPath))` to `while (boost::filesystem::exists(pidfile))`, which never becomes false as the `mongod.lock` file this new version is waiting for is only ever emptied, not deleted.

This fixes that to wait for the file to be either deleted *or* emptied.