SERVER-59719 Ensure resharding commit/abort completes before verifying
ShardsvrCommitReshardCollectionCommand's commit() call and
ShardsvrAbortReshardCollectionCommand's abort() call each perform a
write which will trigger the state document to be deleted on the donor
and recipient. To verify this is done, those commands perform a read to
check if the state documents still exist. Now that the RSTL is not
acquired during reads, it is not guaranteed that the command's opCtx
will be interrupted when performing a read despite calling
setAlwaysInterruptAtStepDownOrUp(). As a consequence, it is possible for
the command's write to have been interrupted due to a step down on the
donor/recipient, causing the document to still exist during the
verification read and triggering a uassert. To resolve this issue, the
commands now do a no-op write before the verification read, ensuring
that the first write has indeed completed without being interrupted.

(cherry picked from commit cca75006b85690faa641a15dfc9940d2a2add52d)