SERVER-61976 Clarify semantics of step up shard version refresh retries
The previous commit for SERVER-61976 added an AsyncTry to retry failed
shard version refreshes on step up. That AsyncTry was using the
cancellation token from an opCtx that goes out of scope shortly after
the AsyncTry is created, and therefore will never be cancelled. Since
the AsyncTry could functionally never be cancelled anyway, this commit
makes that clear by using CancellationToken::uncancelable() instead.