SERVER-61444 Fix race in resharding coordinator service unit test
Changes in the previous commit for SERVER-61444 refactored
resharding_coordinator_service_test.cpp to use functions for behavior
common between test cases. In doing so, the creation order for the
ReshardingCoordinatorService and the PauseDuringStateTransitions guard
was reversed. If the guard is created after the service, it is possible
that the service progresses through a state transition without the guard
having observed it. If this happens, when waiting for that state, the
guard will hang indefinitely instead of returning immediately. The guard
now is once again created before the service to resolve this issue.