SERVER-60321 Fix edge cases when setting up state for Lock-free reads
When LFR is setup we should now always:

* Open a storage snapshot
* Stash the catalog
* Mark the OperationContext that we are in LFR
None of the three is done if we are not LFR.

This fixes an edge case where the opCtx was marked as LFR when no
collection was found and no snapshot or catalog was stashed. A sub op
could then stash a catalog that wasn't cleaned up properly when tearing
down. Anything that happened afterwords, like the profile logic, now had
a stashed catalog when there should be none.

Also fixed when we lock for reading on a view that we are not marking
the OperationContext to be LFR as we never opened a storage snapshot or
stashed the catalog.