SERVER-65201 Log compact form of ignored errors in ReshardingTest.
Moves the check for whether the resharding operation run by the
background thread as part of withReshardingInBackground() succeeded or
failed to the main thread. This enables the main thread to selectively
choose how to format the command response when the resharding operation
is intentionally interrupted due to a separate JavaScript error.