SERVER-67305 Fix lock-free read timestamp order during repl stepup
Eliminate the race between setting read source to lastApplied and getting
the lastApplied timestamp for storage snapshot. When transitioning out of
primary catchup mode we may get a snapshot at lastApplied concurrent with
out-of-order writers.

Changed so the lastApplied timestamp is determined when calling
setTimestampReadSource to allow us to re-verify that we need to read
at lastApplied before opening the storage snapshot.