SERVER-52905 Complete TODO listed in SERVER-49834
There are two TODO's on SERVER-49834 in jstests/replsets/tenant_migration_concurrent_writes_on_donor_util.js: one for unskipping the captrunc command and one for unskipping the compact command.

However, SERVER-49834 is about fsyncLock and fsyncUnlock, not captrunc or compact.

The TODO for captrunc was probably meant to be on SERVER-52895, where we ultimately decided not to support captrunc or emptycapped in serverless. We later banned capped collections entirely in serverless in CLOUDP-106443. So, I just removed all testing for capped collections from this file.

The TODO for compact was probably meant to be on SERVER-52896, where we ultimately decided not to support compact in serverless. So, I similarly removed testing for compact from this file.