SERVER-78409 Erase from outstanding queries list by query in mongo client
Removing from the list by iterator queryIter resulted in failure because
queryIter can be invalidated by the time the lambda function is called and
calls the copy constructor of the queryIter (since we don't hold the lock).
Here we remove the query from the list by traversing the list, which is not
expected to be very large.