SERVER-79853 [CQF] Correctly handle hashed shard keys when performing shard filtering
When a shard key contains a hashed component, the ShardFilterer expects
us to pass it a value which is also hashed. We achieve this by inserting
an invocation of the 'shardHash' SBE builtin function where necessary.

This patch also updates the CQF fallback mechanism to enable CQF for
hashed shard keys when a natural hint is provided and/or there is no
hashed index backing the shard key (this is allowed by SERVER-78253).