SERVER-65534 Introduce host requirement kRunOnceAnyNode for $documents
Running db.aggregate([{$documents: [{}]}, {$out: "test"}]) on mongos will fail because the host requirement of $queue kLocalOnly is translated to kMongoS, which conflicts with the requirement of $out which needs kPrimaryShard. In other words, kLocalOnly and kPrimaryShard cannot coexist in the same pipeline.

To avoid the conflict, this PR introduces a new requirement HostTypeRequirement::kRunOnceAnyNode which works well with kPrimaryShard.