SERVER-81295 Fix v1/v2 change stream resume token transition
Change stream implementation assumes change events are totally
ordered by event '_id', which also serves as a resume token.
In server versions prior to v7.0 the default resume token version
is v1. Starting from v7.0 the default resume token version is v2,
but the server still supports clients supplying v1 resume tokens
(e.g., after server upgrade).

The token version is a part of a token and all v1 tokens sort before
v2 tokens. To ensure the correct event ordering, the server from v7.0
generates v1 tokens to be compared with the client's v1 token until
the referred event is either located or surpassed.

In this commit we fix the following issues:
 - When resuming from the event within a transaction we need to
consider 'txnOpIndex', because the whole transaction gets the same
timestamp.
 - Splitting large events is only supported with v2 tokens. If a
resume token is a high-watermark v1 token, the next event sharing
the same timestamp will have token version v1. If this event is
over 16MB, splitting it will fail because of the token version. To
fix this, we convert the client's high-watermark v1 token to v2.