SERVER-82717 Fix invalid assertion in query planner removeIndexRelevantTag
One phase of query planning attaches RelevantTag objects to predicates
which may be indexable, then a later phase removes some of them.

A fuzzer found a case, involving a rarely(?) used internal operator
$_internalSchemaAllElemMatchFromIndex, where we incorrectly attempt to
remove a tag that was never added.

It seems simpler to just leave the predicate alone when the tag isn't
present--there's nothing to remove, so we've succeeded in removing it.
It's unclear why this was an assertion in the first place.

GitOrigin-RevId: 40f3f09606de8e41cf19b13819dc58a60811c1e7