SERVER-84013 Avoid invoking MatchExpression::optimize() on children of $or in the subplanner.
The subplanner invokes the query planner for each branch of a rooted
$or. Doing this requires constructing a CanonicalQuery representing a
single branch of the $or. A side-effect of cosntructing a CanonicalQuery
is that the MatchExpression is optimized. The subplanner assumes that
both of these MatchExpressions are identical and makes decisions about
index bounds based on the order of predicates in the MatchExpression.
There are queries (as demostrated in the regression test this patch
introduces) for which this assumption is not true and leads to
incorrect index bounds and thus results.

This patch fixes the bug by ensuring that the subplanner does not
invoke MatchExpression optimization a second time when constructing a
copy of the children of the $or.

GitOrigin-RevId: 854ce65ffc67bbefdcfba7d7286a44e8ea5ad1a6