SERVER-78803 Reject concurrent query settings updates in sharded clusters (#17060)
SERVER-78803 Reject concurrent query settings updates in sharded
clusters

Until now, a new value of a cluster parameter was just replacing the old
one.
For the query settings project it is important to read / modify / write
cluster parameter values. In this scenario, concurrent calls of
setQuerySettings command may lead to seemingly successful, but incorrect
results.
This commit introduces an additional 'previousTime' parameter to ensure
that the cluster parameter value that was read has not changed at the
time
of writing, otherwise the operation must fail.
SERVER-78802 has added the 'previousTime' parameter and the related
check
for replica set deployments. This change is adding the check to sharded
cluster
deployments.
As of now, the optional 'previousTime' parameter is only being used by
the
query settings command. The design consideration for this new parameter
and
the related check are described in the design document attached to the
parent
ticket SERVER-78801.

GitOrigin-RevId: f5f35648c676f7dd132ba69d34bbd5251b71a2e1