SERVER-83757 [CQF] Draw one sample to estimate all predicates in a query
When a query has multiple predicates, and the sampling CE method is
used, we run a sampling query for each predicate separately. The
sampling query is responsible for both 1. choosing a random sample of
documents 2. counting the fraction of sampled documents that match
the predicate.

Using a separate sample for each predicate can lead to "contradictory"
estimates. For example, you could estimate that {c: "US", s: "NY"}
matches more documents than {c: "US"} -- this is a bad estimate because
adding the {s: "NY"} filter cannot possibly *increase* the number of
matching documents.

This commit prevents these contradictory estimates by (conceptually)
drawing a single sample and reusing it for all predicates in the query.

Physically, we materialize only a handful of record IDs: one per chunk,
controlled by 'internalCascadesOptimizerSampleChunks'.

GitOrigin-RevId: 36b07d142f0cf65d4b554160d67e152c64964262