SERVER-83610: Reduce required privileges for $documents by marking it… (#18702)
[SERVER-83610](https://jira.mongodb.org/browse/SERVER-83610): Reduce
required privileges for $documents by marking it as an initial source

Aggregations starting with $documents do not read from a collection;
they produce the data specified in the $documents stage. So, these
queries should not require any particular namespace privileges aside
from those required by other document sources in the pipeline. This
change makes it so that $documents is marked as an "initial source" to
indicate that it does not require any particular namespace privileges.

GitOrigin-RevId: fa0030e240db99549a3d57fc03e60065a6b31252