SERVER-86415 Make BSONColumnBuilder safe to move
Internally stored callbacks capturing pointers have been removed.

Instead callbacks are constructed when needed and passed through the
interfaces so they do not need to be stored.

To avoid overhead of constructing large std::function, the callbacks are
now passed as template parameters.

As a secondary benefit the size of the BSONColumnBuilder is around 200
bytes less.

GitOrigin-RevId: 53c48cdba4f400b87b3f15754c4e1e7e1a9a3805