SERVER-85707 Propagate $_generateV2ResumeTokens to config shard change stream (#18383)
SERVER-85707 Propagate $_generateV2ResumeTokens to config shard

We introduced the $_generateV2ResumeTokens internal aggregate command
parameter in v6.0 for test purposes. When passed through the mongoS
connection, it should be set in all internal cluster requests
effectively changing the default resume token version for the current
request in the whole cluster. Before this change, the
$_generateV2ResumeTokens parameter value for the internal request to the
config server relied on the expression context object, which can also be
modified by the new $changeStreamSplitLargeEvent stage. With this
change, we read the parameter value from the original aggregate command
request and so remove a possible conflict with the
$changeStreamSplitLargeEvent stage.

GitOrigin-RevId: db807e6243eabfb2097948bf6a5f0a1ca5b0c070