SERVER-85876: Reject requests matching QuerySettings with reject=true (#18645)
SERVER-85876
Reject requests matching QuerySettings with reject=true

If an admin has set reject in QuerySettings matching an incoming
find/agg/distinct, the request should be failed.

However, explaining such a query is comparatively cheap and is useful
as a debugging tool. As such, explains for such queries should still be
permitted.

GitOrigin-RevId: ca54838c041844066760e2dadf682e6a05679260