SERVER-87976 Fix checking for new elements during transition in HashRoaringSet (#20359)
SERVER-87976 Fix checking for new elements during transition in
HashRoaringSet.

During the transition of elements from the hash table to Roaring
Bitmaps, we should first check if the element being added is present in
the hash table. If it's not found there, then try to add it to the
Roaring Bitmaps.

GitOrigin-RevId: 235192b0014fa966d50d16559220d26f02d97744