SERVER-87052: Prevent rejection of system and administrative queries via query settings (#19937)
SERVER-87052

Prevent rejection of system and administrative queries via query
settings

Some queries are "administrative", and it would not be beneficial to
allow them to be rejected.
To avoid users accidentally creating a state which is hard to reverse or
diagnose, prevent query settings being used to reject selected commands.

GitOrigin-RevId: ee9746446009c7fbfa459ac3bf37a5128f13d467