SERVER-85892 Fix incorrect output namespace in $merge spec (#19846)
SERVER-85892 Fix incorrect output namespace in $merge spec

During the pipeline validation, the original implementation passes in
the namespace 'pExpCtx->ns' to validate the pipeline in whenMatched.
However, this is incorrect as that the pipeline is operated on the
collection specified in 'into' rather than the db/collection the
top-level pipeline operates on. This bug is revealed when a command
operates on a collection-less pipeline (i.e. 'pExpCtx->ns' points to
$cmd.aggregate collection) when query stats enabled or under sharded
cluster.

This commit corrects the namespace for pipeline validation and adds
jstest to verify correctness under various environments.

GitOrigin-RevId: d061e83f3742300e9492bf0531de3cd7b7de0efc