SERVER-89011 Remove attempts to get Bonsai plan executor
Delete noPassthrough/ tests which get CQF feature flag that no longer
work. After this change, the CQF feature flag will no longer work. Since
we are committing this directly on the v8.0 branch (and not on master,
which at the time of writing is v8.1), we aren't removing Bonsai functionality.

Our analysis shows these checks have a small but measurable overhead for
short running queries like those in YCSB.

GitOrigin-RevId: 789198469ae3eb88329f5e80745a5a821ef615ac