SERVER-89268: Serialize 'let' variables by wrapping in $literal, serializing missing 'let' variables with $$REMOVE (#21227)
SERVER-89268: Serialize 'let' variables by wrapping in $literal,
serializing missing 'let' variables with $$REMOVE

When serializing values, for example to populate the 'let' parameter of
a command to be sent over the wire, we need to be careful. First,
missing values should be serialized as "$$REMOVE", otherwise they might
be incorrectly omitted or serialized as empty objects ({}). We should
wrap other kinds of values in $literal to avoid a scenario like the
following: suppose we had a user-defined 'let' specified as {let: {a:
{$literal: "$notAFieldName"}}}. On mongos, this will be evaluated to the
string "$notAFieldName". When we serialize it again for the shard
commands, it must appear as {$literal: "$notAFieldName"}, not simply
"$notAFieldName", since the latter will be treated as a field name by
mongods.
GitOrigin-RevId: 5695070e5776cfa4de7c68a99d07e0a3783c63fc