SERVER-90006 Fix up commit message validation logic (master) (#22052)
(This is a copy of https://github.com/10gen/mongo/pull/21906, applied to
master branch)

Simplifies commit message validation to the simplest possible logic: a
straightforward regex on the PR title.

Note: This PR also includes the addition of a `github_pr_alias`, so that
the new GitHub merge queue uses the correct set of tests on push.
GitOrigin-RevId: 4d30700235735798a52a344b291f8ca0bf530cb0