SERVER-90117 Make collMod not ignore participants' responses from shards not owning chunks (#22365)
# Issue

SERVER-90117

# Description

Today, the collMod coordinator broadcasts `shardsvrCollModParticipant`
to all shards in the cluster. The coordinator will parse each shard
response depending if that shard owns any chunk.

This ticket updates that behavior to not ignore those responses and
retry retriable DDL errors from shards not owning chunks. On the other
hand, if the error is not retriable, the collMod coordinator is going to
swallow it as before.

GitOrigin-RevId: c0fc4627b98a348d13863628d0c51aca17e1ab5b