SERVER-91494: Avoid past-the-end access in DocumentSourceBucket (#23578)
DocumentSourceBucket::createFromBson checks the first character of the
string given for groupBy, but does not ensure that the string is not
empty.

Anything in this description will be included in the commit message.
Replace or delete this text before merging. Add links to testing in the
comments of the PR.

----

Original PR: #23495

----

[evergreen](https://spruce.mongodb.com/version/66705050ef83b80007ec468a)
green.

GitOrigin-RevId: 2034d734099733f2e142a96bf4f7f38eb11c3fb6