SERVER-91258 Increase oplog size in change_streams_split_event tests (#23675)
The default oplog size in deployments created with ShardingTest is 16MB,
which is not enough to accommodate 4 nearly 16-MB large changes (2
updates and 2 replaces) created by this test. As a result, the oplog
might get truncated during the test which leads to sporadic test
failures. Therefore, we need to increase the oplog size.

GitOrigin-RevId: f9d781c5c6aeb92a6e06a7d02fc02d71c64cb0d3