SERVER-91512 Preserve unowned values in BlockToRow::doSaveState (#24083)
This PR address a problem in `BlockToRow::doSaveState` which fails to
preserve unowned values in a block when there are still some values not
yet deblocked.

The fix is to ignore the status of `slotsAccessible()` in
`BlockToRow::doSaveState` to ensure that any unowned values remaining in
the block are preserved.

In order to reproduce the issue in unit test, a new SBE stage is
introduced -- VirtualScanStage. This stage mimics the behavior in a
collection scan which releases memory from the cursor in getNext() and
saveState(). With VirtualScanStage, we are able to reproduce the failure
in the added unit test `BlockToRowNoValuesFilteredObjects`. The issue
was also reproduced in a JS integration test,
`timeseries_cursor_batches.js`, which runs a multi-batch aggregation
over a single block of documents.

Co-authored-by: Chi-I Huang <chii.huang@mongodb.com>
Co-authored-by: Hana Pearlman <hana.pearlman@mongodb.com>

---------

Co-authored-by: Chi-I Huang <8468853+chiihuang@users.noreply.github.com>
GitOrigin-RevId: 6ee54a295e81363bf9e3a6e6413f1e207fca8c8c