SERVER-91406 Allow $match to move ahead of mongoS change stream stages (#24200)
This change allows $match to always move ahead of
$_internalChangeStreamEnsureResumeTokenPresent and
$_internalChangeStreamHandleTopologyChange stages. This is a
precondition for the $changeStreamSplitLargeEvent stage. It also will
move the problematic $match stages to the shard pipelines improving the
overall change stream performance. Before this change some $match
expression (e.g., $jsonSchema) would not move ahead of the
aforementioned stages.

Anything in this description will be included in the commit message.
Replace or delete this text before merging. Add links to testing in the
comments of the PR.

GitOrigin-RevId: cbfd00d3b0cae2a1e195f3e2b4e8fb2f2aa9c1ed