SERVER-91818 More forcefully step down the config shard's primary to v7.3 (#24528)
[SERVER-91818](https://jira.mongodb.org/browse/SERVER-91818)

This is the backport of https://github.com/10gen/mongo/pull/24029 to
v7.3.

On slow builds like enterprise-rhel80-debug-tsan (and other rhel80-debug
and windows-debug builds), a data-bearing config server can fail to
elect a primary because secondaries don't replicate fast enough to catch
up between writes. Instead of asking one of the secondaries to step up,
instead ask the primary to step down, which will trigger an election.

GitOrigin-RevId: 385d99db0b76dc2bf1cdb08110531f6a1672421d