SERVER-92417 Enable indexed plans for expressions referencing system … (#24744)
…variables (#23656)

This commit is meant to achieve the effects of SERVER-83524 but on the
7.3 branch and basically is a frontport of SERVER-91625, which has done
the same changes for 7.0. Since the 7.0/7.3 branches still have the SBE
plan cache enabled, we cannot just re-enable the constant folding of
ExpressionFieldPath. This commit disables queries with expressions
referencing system variables from running in SBE. This allows us to
safely re-enable the constant folding optimization since queries will
use the classic engine and thus classic plan cache as well, avoiding the
problems encountered in SERVER-74264.

Anything in this description will be included in the commit message.
Replace or delete this text before merging. Add links to testing in the
comments of the PR.

🌲

[patch](https://spruce.mongodb.com/version/66952a8e69b85a000755add0/tasks?sorts=STATUS%3AASC%3BBASE_STATUS%3ADESC)

(cherry picked from commit 019517b0a3b534288eeda23068c8af850760428c)

Co-authored-by: Ben Shteinfeld <ben.shteinfeld@mongodb.com>
GitOrigin-RevId: 4b0214e317107d421bdf9379cb1ef8dcb61d5d99