SERVER-93551: Rebalance file/task count for update fuzzers (#26152) (#26517)
The fuzzer currently tests 400 files in 20 tasks. This can timeout when
testing with 7.0 during multiversion tasks. This PR changes it to 400
files in 40 tasks, so that each task has half as many files and should
no longer time out.

Anything in this description will be included in the commit message.
Replace or delete this text before merging. Add links to testing in the
comments of the PR.

GitOrigin-RevId: eecc1ece8520103f7c99dd4466992745b2054632