evergreen: y2030 - stop producing SHA1 and MD5 files, and only provide SHA256 files
To comply with complete withdrawal of SHA1 usage by both IETF RFC
(non-fips) and NIST CMPV (fips) by 2030, stop producing SHA1 and MD5
checksum files in favor of only providing SHA256 checksum files.

As we enter 2026, software produced today might be in active use in
2030 hence it is best to be forward looking with this.

If anything breaks, this can be reverted with a migration plan figured
out to be executed by 2030.